num: Move user-public float parsing items directly under core::num
Follow up to the previous commit refactoring `core::num` by moving the
user-facing error types and trait implementations back out of `imp` and
into a new module under `core::num`.
