+++ tests/codegen-llvm/overflow-checks.rs
@@ -21,7 +21,7 @@ pub unsafe fn add(a: u8, b: u8) -> u8 {
     //        CHECK: i8 noundef{{( zeroext)?}} %a, i8 noundef{{( zeroext)?}} %b
     //        CHECK: add i8 %b, %a
     //        DEBUG: icmp ult i8 [[zero:[^,]+]], %a
-    //        DEBUG: call core::num::overflow_panic::add
+    //        DEBUG: call core::num::imp::overflow_panic::add
     //        DEBUG: unreachable
     // NOCHECKS-NOT: unreachable
     //     NOCHECKS: ret i8 %0
